create model_timestamp to avoid circular imports
